+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### [0.14.2]
+
 ### Added
 
 - New value added to git `ChangeType` enum: `delete, sourceRename`
@@ -355,7 +357,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[Unreleased]: https://github.com/microsoft/azure-devops-rust-api/compare/0.14.1...HEAD
+[Unreleased]: https://github.com/microsoft/azure-devops-rust-api/compare/0.14.2...HEAD
+[0.14.2]: https://github.com/microsoft/azure-devops-rust-api/compare/0.14.1...0.14.2
 [0.14.1]: https://github.com/microsoft/azure-devops-rust-api/compare/0.14.0...0.14.1
 [0.14.0]: https://github.com/microsoft/azure-devops-rust-api/compare/0.13.0...0.14.0
 [0.13.0]: https://github.com/microsoft/azure-devops-rust-api/compare/0.12.0...0.13.0